Move class GenericType into gtfs package
